// src/Type/excel.type.ts
export interface CountryData {
    country: string;
    year: number;
    population: number;
    continent?: string; // อาจมีหรือไม่มีก็ได้
  }
  
  export interface ContinentVisibility {
    Asia: boolean;
    Europe: boolean;
    America: boolean;
    Africa: boolean;
    Oceania: boolean;
  }
  

  